Rename SearchFilters -> SearchFiltersPrimary
