+++ src/components/index.js
@@ -133,8 +133,8 @@ export { default as Page } from './Page/Page';
 export { default as PriceFilter } from './PriceFilter/PriceFilter';
 export { default as Reviews } from './Reviews/Reviews';
 export { default as SavedCardDetails } from './SavedCardDetails/SavedCardDetails';
-export { default as SearchFilters } from './SearchFilters/SearchFilters';
 export { default as SearchFiltersMobile } from './SearchFiltersMobile/SearchFiltersMobile';
+export { default as SearchFiltersPrimary } from './SearchFiltersPrimary/SearchFiltersPrimary';
 export { default as SearchFiltersSecondary } from './SearchFiltersSecondary/SearchFiltersSecondary';
 export { default as SearchMap } from './SearchMap/SearchMap';
 export { default as SearchMapGroupLabel } from './SearchMapGroupLabel/SearchMapGroupLabel';
