+++ src/containers/SearchPage/SearchPage.css
@@ -74,7 +74,7 @@
   white-space: nowrap;
 }
 
-.searchFilters {
+.searchFiltersPrimary {
   display: none;
 
   @media (--viewportMedium) {
