+++ src/translations/de.json
@@ -701,17 +701,6 @@
   "SavedCardDetails.replaceCardTitle": "Andere Zahlungsmethode verwenden",
   "SavedCardDetails.savedPaymentMethodPlaceholderDesktop": "•••• •••• •••• {last4Digits}",
   "SavedCardDetails.savedPaymentMethodPlaceholderMobile": "…{last4Digits}",
-  "SearchFilters.amenitiesLabel": "Ausstattung",
-  "SearchFilters.categoryLabel": "Kategorie",
-  "SearchFilters.filtersButtonLabel": "Filter",
-  "SearchFilters.foundResults": "{count, number} {count, plural, one {Ergebnis} other {Ergebnisse}}",
-  "SearchFilters.keywordLabel": "Schlüsselwort",
-  "SearchFilters.loadingResults": "Lade Suchergebnisse…",
-  "SearchFilters.loadingResultsMobile": "Lade…",
-  "SearchFilters.moreFiltersButton": "{count, plural, =0 {Andere Filter} other {Andere Filter • #}}",
-  "SearchFilters.noResults": "Keine Einträge gefunden.",
-  "SearchFilters.noResultsMobile": "Keine Ergebnisse.",
-  "SearchFilters.openMapView": "Karte",
   "SearchFiltersMobile.amenitiesLabel": "Ausstattung",
   "SearchFiltersMobile.cancel": "ABBRECHEN",
   "SearchFiltersMobile.categoryLabel": "Kategorie",
@@ -719,6 +708,10 @@
   "SearchFiltersMobile.keywordLabel": "Schlüsselwort",
   "SearchFiltersMobile.resetAll": "Alles zurücksetzen",
   "SearchFiltersMobile.showListings": "Zeige {count, number} {count, plural, one {Sauna} other {Saunas}}",
+  "SearchFiltersPrimary.foundResults": "{count, number} {count, plural, one {Ergebnis} other {Ergebnisse}}",
+  "SearchFiltersPrimary.loadingResults": "Lade Suchergebnisse…",
+  "SearchFiltersPrimary.moreFiltersButton": "{count, plural, =0 {Andere Filter} other {Andere Filter • #}}",
+  "SearchFiltersPrimary.noResults": "Keine Einträge gefunden.",
   "SearchFiltersSecondary.apply": "Anwenden",
   "SearchFiltersSecondary.cancel": "Abbrechen",
   "SearchFiltersSecondary.resetAll": "Alles zurücksetzen",
