+++ src/translations/en.json
@@ -697,15 +697,6 @@
   "SavedCardDetails.replaceCardTitle": "Use another payment method",
   "SavedCardDetails.savedPaymentMethodPlaceholderDesktop": "•••• •••• •••• {last4Digits}",
   "SavedCardDetails.savedPaymentMethodPlaceholderMobile": "…{last4Digits}",
-  "SearchFilters.filtersButtonLabel": "Filters",
-  "SearchFilters.foundResults": "{count, number} {count, plural, one {result} other {results}}",
-  "SearchFilters.keywordLabel": "Keyword",
-  "SearchFilters.loadingResults": "Loading search results…",
-  "SearchFilters.loadingResultsMobile": "Loading…",
-  "SearchFilters.moreFiltersButton": "{count, plural, =0 {More filters} other {More filters • #}}",
-  "SearchFilters.noResults": "Could not find any listings.",
-  "SearchFilters.noResultsMobile": "No results.",
-  "SearchFilters.openMapView": "Map",
   "SearchFiltersMobile.amenitiesLabel": "Amenities",
   "SearchFiltersMobile.cancel": "CANCEL",
   "SearchFiltersMobile.categoryLabel": "Category",
@@ -713,6 +704,10 @@
   "SearchFiltersMobile.keywordLabel": "Keyword",
   "SearchFiltersMobile.resetAll": "Reset all",
   "SearchFiltersMobile.showListings": "Show {count, number} {count, plural, one {sauna} other {saunas}}",
+  "SearchFiltersPrimary.foundResults": "{count, number} {count, plural, one {result} other {results}}",
+  "SearchFiltersPrimary.loadingResults": "Loading search results…",
+  "SearchFiltersPrimary.moreFiltersButton": "{count, plural, =0 {More filters} other {More filters • #}}",
+  "SearchFiltersPrimary.noResults": "Could not find any listings.",
   "SearchFiltersSecondary.apply": "Apply",
   "SearchFiltersSecondary.cancel": "Cancel",
   "SearchFiltersSecondary.resetAll": "Reset all",
