+++ src/translations/es.json
@@ -701,17 +701,6 @@
   "SavedCardDetails.replaceCardTitle": "Usar otro método de pago",
   "SavedCardDetails.savedPaymentMethodPlaceholderDesktop": "•••• •••• •••• {last4Digits}",
   "SavedCardDetails.savedPaymentMethodPlaceholderMobile": "…{last4Digits}",
-  "SearchFilters.amenitiesLabel": "Servicios",
-  "SearchFilters.categoryLabel": "Categoría",
-  "SearchFilters.filtersButtonLabel": "Filtros",
-  "SearchFilters.foundResults": "{count, number} {count, plural, one {resultado} other {resultados}}",
-  "SearchFilters.keywordLabel": "Palabra clave",
-  "SearchFilters.loadingResults": "Buscando resultados…",
-  "SearchFilters.loadingResultsMobile": "Buscando…",
-  "SearchFilters.moreFiltersButton": "{count, plural, =0 {Más filtros} other {Más filtros • #}}",
-  "SearchFilters.noResults": "No se han encontrado resultados.",
-  "SearchFilters.noResultsMobile": "No hay resultados.",
-  "SearchFilters.openMapView": "Mapa",
   "SearchFiltersMobile.amenitiesLabel": "Servicios",
   "SearchFiltersMobile.cancel": "CANCELAR",
   "SearchFiltersMobile.categoryLabel": "Categoría",
@@ -719,6 +708,10 @@
   "SearchFiltersMobile.keywordLabel": "Palabra clave",
   "SearchFiltersMobile.resetAll": "Borrar filtros",
   "SearchFiltersMobile.showListings": "Mostrar {count, number} {count, plural, one {sauna} other {saunas}}",
+  "SearchFiltersPrimary.foundResults": "{count, number} {count, plural, one {resultado} other {resultados}}",
+  "SearchFiltersPrimary.loadingResults": "Buscando resultados…",
+  "SearchFiltersPrimary.moreFiltersButton": "{count, plural, =0 {Más filtros} other {Más filtros • #}}",
+  "SearchFiltersPrimary.noResults": "No se han encontrado resultados.",
   "SearchFiltersSecondary.apply": "Aplicar",
   "SearchFiltersSecondary.cancel": "Cancelar",
   "SearchFiltersSecondary.resetAll": "Borrar",
