+++ src/translations/fr.json
@@ -695,17 +695,6 @@
   "SavedCardDetails.replaceCardTitle": "Utiliser un autre moyen de paiement",
   "SavedCardDetails.savedPaymentMethodPlaceholderDesktop": "•••• •••• •••• {last4Digits}",
   "SavedCardDetails.savedPaymentMethodPlaceholderMobile": "…{last4Digits}",
-  "SearchFilters.amenitiesLabel": "Équipements",
-  "SearchFilters.categoryLabel": "Catégorie",
-  "SearchFilters.filtersButtonLabel": "Filtres",
-  "SearchFilters.foundResults": "{count, number} {count, plural, one {résultat} other {résultats}}",
-  "SearchFilters.keywordLabel": "Mot-clé",
-  "SearchFilters.loadingResults": "Chargement des résultats…",
-  "SearchFilters.loadingResultsMobile": "Chargement…",
-  "SearchFilters.moreFiltersButton": "{count, plural, =0 {Plus de filtres} other {Plus de filtres • #}}",
-  "SearchFilters.noResults": "Aucune annonce trouvée.",
-  "SearchFilters.noResultsMobile": "Aucun résultat.",
-  "SearchFilters.openMapView": "Carte",
   "SearchFiltersMobile.amenitiesLabel": "Équipements",
   "SearchFiltersMobile.cancel": "ANNULER",
   "SearchFiltersMobile.categoryLabel": "Catégorie",
@@ -713,6 +702,10 @@
   "SearchFiltersMobile.keywordLabel": "Mot-clé",
   "SearchFiltersMobile.resetAll": "Remettre à zéro",
   "SearchFiltersMobile.showListings": "Voir {count, number} {count, plural, one {sauna} other {saunas}}",
+  "SearchFiltersPrimary.foundResults": "{count, number} {count, plural, one {résultat} other {résultats}}",
+  "SearchFiltersPrimary.loadingResults": "Chargement des résultats…",
+  "SearchFiltersPrimary.moreFiltersButton": "{count, plural, =0 {Plus de filtres} other {Plus de filtres • #}}",
+  "SearchFiltersPrimary.noResults": "Aucune annonce trouvée.",
   "SearchFiltersSecondary.apply": "Valider",
   "SearchFiltersSecondary.cancel": "Annuler",
   "SearchFiltersSecondary.resetAll": "Effacer",
